CreateRecipe Back End Connection
The create recipe form is now connected to the backend nodejs end point. The end point will receive the title, steps, and ingredient list.
TODO:
Want to make it so as the user types their recipe information, there is a layout that updates in real time showing them their recipe.
Want to add a toggle option to the recipe steps so they can enter their steps like the ingrient list is.